checkpoint: dialog and components separated
