+++ src/components/SelectTopic.vue
@@ -1,0 +1,42 @@
+
+<script setup lang="ts">
+import { computed, ref } from 'vue';
+import { Topic } from '../types/topic';
+
+const props = defineProps<{
+  topics: Topic[]
+  submit: (topic: Topic) => Promise<void>
+}>();
+
+const searchQuery = ref('');
+
+const filteredTopics = computed(() => {
+	if (!searchQuery.value) return props.topics;
+	return props.topics
+		.filter(topic => {
+			const query = searchQuery.value.toLowerCase();
+			const includesName = topic.name.toLowerCase().includes(query);
+			return includesName;
+		});
+});
+
+const submit = (topic: Topic) => {
+	props.submit(topic);
+};
+
+</script>
+
+<template>
+	<div class="mb-6 flex justify-between items-center">
+		<input type="text" v-model="searchQuery"
+			class="block bg-transparent outline-none border-b border-gray-400 py-1 w-[400px]" placeholder="Search">
+	</div>
+
+	<ul>
+		<li v-for="topic, index in filteredTopics" :key="index" :class="{'border-b': index !== filteredTopics.length - 1}"
+			class="border-white overflow-hidden flex justify-between py-2 px-3 cursor-pointer hover:bg-[#252526]" @click="submit(topic)">
+			<span class="mr-4">{{ topic.name }}</span>
+			<i class="bi-arrow-right text-xl"></i>
+		</li>
+	</ul>
+</template>
